+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## Unreleased
 ### Added
 - Added cider package and CHANGELOG.md
+- New test to check if exists Icons.share
+
+### Changed
+- Putting share function on cnpj\_controller